+++ PRO3/PRO3/InAndOut.hpp
@@ -64,7 +64,7 @@ public:
      ** Post-Conditions: None
      ** Return: none
      *********************************************************************/
-    void saveResult(struct city* tour, int distance, int listLength);
+    void saveResult(struct city* tour, long distance, int listLength);
 
     /*********************************************************************
      ** Function: runtime
